raw conversion of OrderedCallMatching specs
